requirements.txt
pip==22.3
django==4.1.2
tzdata==2022.5
asgiref==3.5.2
sqlparse==0.4.3
setuptools==63.2.0
djangorestframework == 3.14.0 
pytz==2022.5